Add protected controller methods (#48)
* add-protected-routes

* add-protected-routes

+++ src/test/java/com/mangalaxy/mango/controller/JobControllerTest.java
@@ -88,7 +88,7 @@ public class JobControllerTest {
     String expectedTitle = "Java Developer";
 
     LocationRequest locationRequest = LocationRequest.builder()
-        .id(1L)
+        .id((short) 1L)
         .country("UA")
         .city("Kyiv")
         .build();
